# docs/openapi.yaml
openapi: 3.0.1
info:
  title: TrueArena API
  description: "Phase 2-3 slice: phone+OTP auth, groups, rooms. See docs/PROJECT_PLAN.md."
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
tags:
- name: auth
  description: "Phone + OTP. In the local profile, code 000000 always verifies."
paths:
  /api/v1/rooms:
    post:
      tags:
      - rooms
      summary: "Create a room (in a group, or ad-hoc if groupId is omitted)"
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoomView"
  /api/v1/rooms/join:
    post:
      tags:
      - rooms
      summary: Join a room by its 6-char code
      operationId: join
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRoomRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoomView"
  /api/v1/groups:
    get:
      tags:
      - groups
      operationId: mine
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupView"
    post:
      tags:
      - groups
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GroupView"
  /api/v1/groups/{id}/members:
    get:
      tags:
      - groups
      operationId: members
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMemberView"
    post:
      tags:
      - groups
      operationId: addMember
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GroupMemberView"
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Exchange a refresh token for a fresh pair
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenResponse"
      security: []
  /api/v1/auth/otp/verify:
    post:
      tags:
      - auth
      summary: Verify an OTP; creates the account on first use and returns JWTs
      operationId: verifyOtp
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenResponse"
      security: []
  /api/v1/auth/otp/request:
    post:
      tags:
      - auth
      summary: Send an OTP code to a phone number (logged to the console by the dev
        SMS stub)
      operationId: requestOtp
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
        required: true
      responses:
        "202":
          description: Accepted
      security: []
  /api/v1/rooms/{id}:
    get:
      tags:
      - rooms
      operationId: one
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoomView"
  /api/v1/me:
    get:
      tags:
      - user
      summary: The authenticated user
      operationId: me
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserView"
  /api/v1/groups/{id}:
    get:
      tags:
      - groups
      operationId: one_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GroupView"
components:
  schemas:
    CreateRoomRequest:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
    RoomMemberView:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        nickname:
          type: string
        connectionStatus:
          type: string
        ready:
          type: boolean
    RoomView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        groupId:
          type: string
          format: uuid
        hostId:
          type: string
          format: uuid
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/RoomMemberView"
        wsUrl:
          type: string
        sessionToken:
          type: string
    JoinRoomRequest:
      required:
      - code
      type: object
      properties:
        code:
          maxLength: 6
          minLength: 6
          type: string
        nickname:
          maxLength: 24
          minLength: 0
          type: string
    CreateGroupRequest:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 60
          minLength: 0
          type: string
    GroupView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    AddMemberRequest:
      required:
      - userId
      type: object
      properties:
        userId:
          type: string
          format: uuid
    GroupMemberView:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
        joinedAt:
          type: string
          format: date-time
    RefreshRequest:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
        user:
          $ref: "#/components/schemas/UserView"
    UserView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        phone:
          type: string
    OtpVerify:
      required:
      - code
      - phone
      type: object
      properties:
        phone:
          pattern: "\\+?[0-9]{6,15}"
          type: string
        code:
          pattern: "[0-9]{6}"
          type: string
    OtpRequest:
      required:
      - phone
      type: object
      properties:
        phone:
          pattern: "\\+?[0-9]{6,15}"
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT